float to double

+++ src/cluster/AbstractMetric.h
@@ -16,7 +16,7 @@
 class AbstractMetric {
 public:
 
-    virtual float distance(Object &o1, Object &o2, bool bUseIntegratedPrediction = false) = 0;
+    virtual double distance(Object &o1, Object &o2, bool bUseIntegratedPrediction = false) = 0;
     virtual Object* center(list<Object*> &lsObjects) = 0;
     virtual void predictMissingData(DataContainer *pContainer) = 0;
 private:
