+++ src/main/Reader.cpp
@@ -47,7 +47,7 @@ void Reader::fill(DataContainer &container) {
             cout << nCounter << " lines read!" << endl;
     }
     time(&end);
-    float dif = difftime(end, start);
+    double dif = difftime(end, start);
     printf("Done! Processing %i lines took %.2lf seconds\r\n", nCounter, dif);
     _dataFile.close();
 }
@@ -81,7 +81,7 @@ Object* Reader::parseObject(string sRow, int &id) {
     Object *pObj = new Object(nAttributeCount);
     pObj->setActualClass(nActualClass);
     pObj->setId(id);
-    float attrValue = 0;
+    double attrValue = 0;
     int nAttribute = 0;
 
     while (pRow < pEnd) {
@@ -105,7 +105,7 @@ Object* Reader::parseObject(string sRow, int &id) {
     return pObj;
 }
 
-float Reader::getNumber(char **data, int &length, char delimiter) {
+double Reader::getNumber(char **data, int &length, char delimiter) {
     length = 0;
     char *pIndex = *data;
     while (pIndex < *data + strlen(*data)) {
@@ -118,7 +118,7 @@ float Reader::getNumber(char **data, int &length, char delimiter) {
     }
     if (length > 0) {
         (*data)[length] = '\0';
-        float result = atof(*data);
+        double result = atof(*data);
         return result;
     } else
         return 0;
